pyproject.toml
[tool.poetry]
name = "video_summarizer"
version = "0.2.0"
description = "Summarise videos using Gen AI"
authors = ["John Mwangi <mail.mwangi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "3.9.4"
aiosignal = "1.3.1"
altair = "5.2.0"
annotated-types = "0.6.0"
anyio = "3.7.1"
appdirs = "1.4.4"
appnope = "0.1.3"
asttokens = "2.4.1"
async-timeout = "4.0.3"
attrs = "23.1.0"
blinker = "1.7.0"
brotli = "1.1.0"
cachetools = "5.3.2"
certifi = "2023.11.17"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
comm = "0.1.4"
contourpy = "1.2.0"
cryptography = "41.0.7"
cycler = "0.12.1"
dataclasses-json = "0.6.3"
debugpy = "1.8.0"
decorator = "5.1.1"
docker-pycreds = "0.4.0"
et-xmlfile = "1.1.0"
exceptiongroup = "1.2.0"
executing = "2.0.1"
fastapi = "0.109.1"
fonttools = "4.47.0"
frozenlist = "1.4.1"
gitdb = "4.0.11"
gitpython = "3.1.40"
google-api-core = "2.15.0"
google-auth = "2.25.2"
googleapis-common-protos = "1.62.0"
greenlet = "3.0.3"
h11 = "0.14.0"
idna = "3.7"
importlib-metadata = "4.13.0"
importlib-resources = "6.1.1"
iniconfig = "2.0.0"
ipykernel = "6.26.0"
ipython = "8.19.0"
ipywidgets = "8.1.1"
jedi = "0.19.1"
jinja2 = "3.1.2"
joblib = "1.3.2"
jsonpatch = "1.33"
jsonpointer = "2.4"
jsonschema = "4.20.0"
jsonschema-specifications = "2023.11.2"
jupyter-client = "8.6.0"
jupyter-core = "5.5.1"
jupyterlab-widgets = "3.0.9"
kiwisolver = "1.4.5"
langchain = "0.0.352"
langchain-community = "0.0.2"
langchain-core = "0.1.3"
langsmith = "0.0.75"
markdown-it-py = "3.0.0"
markupsafe = "2.1.3"
marshmallow = "3.20.1"
matplotlib = "3.8.2"
matplotlib-inline = "0.1.6"
mdurl = "0.1.0"
multidict = "6.0.4"
munkres = "1.1.4"
mypy = "1.8.0"
mypy-extensions = "1.0.0"
nest-asyncio = "1.5.8"
numexpr = "2.8.8"
numpy = "1.26.2"
openai = "0.28.1"
openapi-schema-pydantic = "1.2.4"
openpyxl = "3.1.2"
packaging = "23.2"
pandas = "2.1.4"
pandas-stubs = "2.1.4.231218"
parso = "0.8.3"
pathtools = "0.1.2"
pexpect = "4.8.0"
pickleshare = "0.7.5"
pillow = "10.3.0"
pip = "23.3.2"
pkgutil-resolve-name = "1.3.10"
plotly = "5.18.0"
pluggy = "1.3.0"
prompt-toolkit = "3.0.42"
protobuf = "4.24.4"
psutil = "5.9.7"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pyarrow = "14.0.2"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pydantic = "2.5.3"
pydantic-core = "2.14.6"
pydantic-settings = "2.1.0"
pydeck = "0.8.0b4"
pygments = "2.17.2"
pyopenssl = "23.3.0"
pyparsing = "3.1.1"
pysocks = "1.7.1"
pytest = "7.4.3"
pytest-subtests = "0.11.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
pytz = "2023.3.post1"
pyu2f = "0.1.5"
pyyaml = "6.0.1"
pyzmq = "25.1.2"
referencing = "0.32.0"
regex = "2023.12.25"
requests = "2.31.0"
rich = "13.7.0"
rpds-py = "0.15.2"
rsa = "4.9"
scikit-learn = "1.3.2"
scipy = "1.11.4"
scrapetube = "2.5.1"
sentry-sdk = "1.39.1"
setproctitle = "1.3.3"
setuptools = "68.2.2"
six = "1.16.0"
smmap = "5.0.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.23"
stack-data = "0.6.2"
starlette = "0.35.0"
streamlit = "1.29.0"
tenacity = "8.2.3"
threadpoolctl = "3.2.0"
tiktoken = "0.5.2"
toml = "0.10.2"
tomli = "2.0.1"
toolz = "0.12.0"
tornado = "6.3.3"
tqdm = "4.66.1"
traitlets = "5.14.0"
types-pytz = "2023.3.1.1"
typing-extensions = "4.9.0"
typing-inspect = "0.9.0"
tzdata = "2023.3"
tzlocal = "5.2"
unicodedata2 = "15.1.0"
urllib3 = "2.1.0"
uvicorn = "0.25.0"
validators = "0.22.0"
wandb = "0.16.1"
watchdog = "3.0.0"
wcwidth = "0.2.12"
wheel = "0.42.0"
widgetsnbextension = "4.0.9"
yarl = "1.9.3"
youtube-transcript-api = "0.6.2"
zipp = "3.17.0"
build = "1.0.3"
cachecontrol = "0.13.1"
cleo = "2.1.0"
crashtest = "0.4.1"
distlib = "0.3.8"
dulwich = "0.21.7"
fastjsonschema = "2.19.1"
filelock = "3.13.1"
installer = "0.7.0"
jaraco-classes = "3.3.0"
keyring = "24.3.0"
more-itertools = "10.2.0"
msgpack = "1.0.7"
pkginfo = "1.9.6"
poetry = "1.7.1"
poetry-core = "1.8.1"
poetry-plugin-export = "1.6.0"
pyproject-hooks = "1.0.0"
rapidfuzz = "3.6.1"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
tomlkit = "0.12.3"
trove-classifiers = "2024.1.31"
virtualenv = "20.25.0"
xattr = "0.10.1"
pymongo = "^4.6.3"
black = "^24.3.0"
streamlit-tags = "^1.2.8"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
python-multipart = "^0.0.9"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
